URL Changes & Allow Script
I changed the URL from Incapsula to Imperva as the URL's have now changed. I also created a separate script that is effectively the same as the other block script except that it adds IP's to the exception list instead of the block list. I may modify the script for a user to pick which option and then execute the script instead of having 2 separate scripts.